[#13]Remove: move accessTokenAtom to atoms folder
